requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cffi==1.17.0
chardet==5.2.0
cryptography==43.0.0
distlib==0.3.4
Django==5.1
djangorestframework==3.15.2
filelock==3.7.1
mysql-connector-python==9.0.0
numpy==2.1.1
pandas==2.2.2
pillow==10.4.0
platformdirs==2.5.2
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2022.1
reportlab==4.2.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.15.1
